pyproject.toml
[project]
name = "rpp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "dmgbuild>=1.6.5",
    "fastapi[all]>=0.116.1",
    "loguru>=0.7.3",
    "pyinstaller>=6.14.2",
    "pywebview>=5.4",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-crud-plus>=1.10.0",
]
